% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{cast_map}
\alias{cast_map}
\title{convert a "long" map into a "wide" one}
\usage{
cast_map(df, id, measure, collapse = " /// ")
}
\arguments{
\item{df}{tibble. rows containing NA in any of following variables are removed.}

\item{id}{string. name of a character variable in \code{df}, its elements can only contain \emph{one} value each}

\item{measure}{string. name of a character variable in \code{df}, its elements can only contain \emph{one} value each}

\item{collapse}{string. used as a separator when collapsing multiple values of \code{measure} which map to a single value of \code{id}}
}
\value{
tibble. column names are \code{id} and \code{measure}.
}
\description{
convert a "long" map into a "wide" one
}
\seealso{
Other mapping: 
\code{\link{melt_map}()}
}
\concept{mapping}
