% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archive.R
\name{download_archive}
\alias{download_archive}
\title{Download HGNC dataset}
\usage{
download_archive(
  url = latest_archive_url(),
  path = getwd(),
  filename = basename(url),
  ...
)
}
\arguments{
\item{url}{A character string naming the URL of the HGNC dataset. It defaults
to the latest available archive.}

\item{path}{A character string with the directory path where the downloaded
file is to be saved. By default, this is the current working directory.}

\item{filename}{A character string with the name of the saved file. By
default, this is inferred from the last part of the URL.}

\item{...}{Additional arguments passed on to \code{\link[=download.file]{download.file()}}.}
}
\description{
Download the latest HGNC approved data set.
}
