% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archive.R
\name{download_hgnc_dataset}
\alias{download_hgnc_dataset}
\title{Download the human gene nomenclature dataset}
\usage{
download_hgnc_dataset(
  url = latest_archive_url(),
  path = getwd(),
  filename = basename(url)
)
}
\arguments{
\item{url}{A character string naming the URL of the HGNC dataset. It defaults
to the latest available archive.}

\item{path}{The directory path where the downloaded file is to be saved. By
default, this value is \code{NULL} and the data is imported directly into memory
without saving into disk.}

\item{filename}{A character string with the name of the saved file. By
default, this is inferred from the last part of the URL.}
}
\value{
The path to the saved file.
}
\description{
\code{\link[=download_hgnc_dataset]{download_hgnc_dataset()}} gets the latest HGNC approved data set.
}
