% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get HGNC cache directory}
\usage{
get_cache_dir()
}
\value{
A character vector of length 1 or NULL
}
\description{
Get HGNC cache directory
}
\examples{
get_cache_dir()

}
