% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_hgnc_file}
\alias{get_hgnc_file}
\title{Get HGNC TSV file in use}
\usage{
get_hgnc_file()
}
\value{
A character vector of length 1 or NULL
}
\description{
Get HGNC TSV file in use
}
\examples{
get_hgnc_file()

}
