% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc-package.R
\docType{package}
\name{hgnc-package}
\alias{hgnc}
\alias{hgnc-package}
\title{hgnc: Download and Import HUGO Gene Nomenclature Committee (HGNC) data into R}
\description{
A set of routines to quickly download and import HUGO Gene data into R.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
}

}
\keyword{internal}
