% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_dataset_example.R
\name{hgnc_dataset_example}
\alias{hgnc_dataset_example}
\title{Example HGNC data set}
\usage{
hgnc_dataset_example()
}
\value{
A \link[tibble:tibble-package]{tibble} whose structure is documented
in \code{\link[=import_hgnc_dataset]{import_hgnc_dataset()}}.
}
\description{
\code{\link[=hgnc_dataset_example]{hgnc_dataset_example()}} provides an example HGNC data set. This example
contains only the first 10,000 gene entries of the HGNC data set dated of
2024-07-26 03:11:20.

This is mostly used in example code as it does not require internet
connection.
}
\examples{
hgnc_dataset_example()

}
