% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_and_convert.R
\name{hgnc_join}
\alias{hgnc_join}
\title{Add a column to a data frame from the hgnc dataset}
\usage{
hgnc_join(
  .data,
  by = "hgnc_id",
  column = "symbol",
  one_to_many = FALSE,
  file = get_hgnc_file()
)
}
\arguments{
\item{.data}{The data frame to join with}

\item{by}{The column to join by, as in \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}}

\item{column}{The column to add to the data frame from the HGNC dataset}

\item{one_to_many}{A boolean value indicating whether cases where the 'by' column maps to multiple valuesof the selected column should be returned.
If set to \code{FASLE} (the default), such one-to-many cases will yield NA. If set to \code{TRUE}, the returned column will be a list of vectors of varying length.}

\item{file}{A file or URL of the complete HGNC data set (in TSV format), as used by \code{\link[=import_hgnc_dataset]{import_hgnc_dataset()}}.}
}
\value{
A data frame with the HGNC column added
}
\description{
Add a column to a data frame from the hgnc dataset
}
\examples{
\dontrun{
library(dplyr)
df <- tibble(hgnc_id = c("HGNC:44948", "HGNC:43240", "HGNC:23357", "HGNC:1855", "HGNC:39400"))

df \%>\%
 hgnc_join(by = 'hgnc_id', column = 'symbol') \%>\%
 hgnc_join(by = 'hgnc_id', column = 'entrez_id')
}

}
