% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc.R
\name{id_convert}
\alias{id_convert}
\title{Convert Gene IDs}
\usage{
id_convert(id, from = guess_id_type(id[1]), to = c("Gene Symbol",
  "Entrez Gene ID", "Ensembl Gene ID", "RefSeq ID"))
}
\arguments{
\item{id}{A vector of gene IDs.}

\item{from}{The ID type of input gene IDs.}

\item{to}{The ID type that user wants to convert to.}
}
\description{
It only converts between gene symbol, Entrez IDs, Ensembl IDs 
and RefSeq ids for human.
}
\examples{
id_convert("TP53", to = "Ensembl")
}
