% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hgnc_dataset.R
\name{import_hgnc_dataset}
\alias{import_hgnc_dataset}
\title{Import HGNC data set}
\usage{
import_hgnc_dataset(file = latest_archive_url(), ...)
}
\arguments{
\item{file}{A file or URL of the  complete HGNC data set (in TSV format).}

\item{...}{Additional arguments to be passed to \code{\link[readr:read_delim]{readr::read_tsv()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of the HGNC data set.
}
\description{
This function imports into memory as a tibble, the complete HGNC data set
from a TSV file.
}
