% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_update.R
\name{last_update}
\alias{last_update}
\title{Last update of HGNC data set}
\usage{
last_update()
}
\value{
A POSIXct date-time object.
}
\description{
This function returns the date of the most recent update of the HGNC data
set.
}
