% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_archive.R
\name{latest_archive_url}
\alias{latest_archive_url}
\title{Latest HGNC archive URL}
\usage{
latest_archive_url(type = c("tsv", "json"))
}
\arguments{
\item{type}{The format of the archive: \code{"tsv"} or \code{"json"}.}
}
\value{
A string with the latest HGNC archive URL.
}
\description{
Latest HGNC archive URL
}
\examples{
latest_archive_url()

}
