% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_archive.R
\name{latest_monthly_url}
\alias{latest_monthly_url}
\title{Latest HGNC monthly URL}
\usage{
latest_monthly_url()
}
\value{
A string with the latest HGNC monthly archive URL.
}
\description{
Latest HGNC monthly URL
}
\examples{
latest_monthly_url()

}
