% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_archive.R
\name{latest_quarterly_url}
\alias{latest_quarterly_url}
\title{Latest HGNC quarterly URL}
\usage{
latest_quarterly_url()
}
\value{
A string with the latest HGNC monthly archive URL.
}
\description{
Latest HGNC quarterly URL
}
\examples{
latest_quarterly_url()

}
