% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_archives.R
\name{list_archives}
\alias{list_archives}
\title{List monthly and quarterly archives}
\usage{
list_archives(type = c("tsv", "json"))
}
\arguments{
\item{type}{The format of the archive: \code{"tsv"} or \code{"json"}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of available archives for download.
}
\description{
This function lists the monthly and quarterly archives currently available.
}
