% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_match.R
\name{matches_once}
\alias{matches_once}
\title{Check if an Element Matches Exactly Once}
\usage{
matches_once(x, y)
}
\arguments{
\item{x}{A vector containing the elements to match.}

\item{y}{A vector in which the elements from \code{x} will be matched.}
}
\value{
A logical vector of the same length as \code{x}, where each element is \code{TRUE} if it matches exactly once in \code{y}, and \code{FALSE} otherwise.
}
\description{
This function checks whether a specific element from vector \code{x} appears exactly once in vector \code{y}.
}
\examples{
\dontrun{
x <- c(1, 2, 3)
y <- c(1, 1, 2, 4)
matches_once(x, y)}

}
\keyword{internal}
