% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_match.R
\name{n_match}
\alias{n_match}
\title{Count the Number of Matches for Each Element in a Vector}
\usage{
n_match(x, y)
}
\arguments{
\item{x}{A vector of elements to be matched.}

\item{y}{A vector in which the elements from \code{x} will be matched.}
}
\value{
An integer vector of the same length as \code{x}, where each element indicates the number of times it matches in \code{y}.
}
\description{
This function counts how many times each element of vector \code{x} matches any element in vector \code{y}.
}
\examples{
\dontrun{
x <- c(1, 2, 3)
y <- c(1, 1, 2, 4)
n_match(x, y)}

}
\keyword{internal}
