% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{qualify_map}
\alias{qualify_map}
\title{qualify a map to be "square"}
\usage{
qualify_map(df)
}
\arguments{
\item{df}{tibble.}
}
\value{
tibble.
}
\description{
qualify a map to be umambiguous and effective (no NA)
}
\examples{
tibble::tribble(
    ~from, ~to,
    NA, 1,
    'B', NA,
    NA, NA,
    'C', 3
) \%>\% qualify_map

tibble::tribble(
    ~from, ~to,
    'A', 0.5,
    'A', 1.5,
    'B', 2
) \%>\% qualify_map

}
