% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{use_cache_dir}
\alias{use_cache_dir}
\title{Enable disk-based caching of hgnc results}
\usage{
use_cache_dir(cache_dir = tools::R_user_dir("hgnc", "cache"))
}
\description{
Enable disk-based caching of hgnc results
}
\examples{
use_cache_dir()

}
