% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc-package.R
\docType{package}
\name{hgnc-package}
\alias{hgnc}
\alias{hgnc-package}
\title{hgnc: Import Human Gene Nomenclature}
\description{
A set of routines to quickly download and import the HUGO Gene Nomenclature Committee (HGNC) data set on mapping of gene symbols to gene entries in other genomic databases or resources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patterninstitute/hgnc}
  \item \url{https://www.pattern.institute/hgnc/}
  \item Report bugs at \url{https://github.com/patterninstitute/hgnc/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte@pattern.institute} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Jacob Munro \email{jacob.e.munro@gmail.com} (\href{https://orcid.org/0000-0002-2751-0989}{ORCID})
}

Other contributors:
\itemize{
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID}) [contributor]
  \item Pattern Institute (04jrgd746) [copyright holder, funder]
}

}
\keyword{internal}
