% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc-package.R
\docType{package}
\name{hgnc-package}
\alias{hgnc}
\alias{hgnc-package}
\title{hgnc: Download and Import the HUGO Gene Nomenclature Committee ('HGNC') Data Set into R}
\description{
A set of routines to quickly download and import the 'HGNC' data set on mapping of gene symbols to gene entries in other popular databases or resources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/hgnc}
  \item \url{https://rmagno.eu/hgnc/}
  \item Report bugs at \url{https://github.com/ramiromagno/hgnc/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
