% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_archive.R
\name{latest_archive_url}
\alias{latest_archive_url}
\title{Latest HGNC archive URL}
\usage{
latest_archive_url(type = c("tsv", "json"))
}
\arguments{
\item{type}{The format of the archive: `"tsv"` or `"json"`.}
}
\description{
Latest HGNC archive URL
}
